feature: allow users to opt out of cluster scoped RBAC
Goldpinger optionally searches for Pods in every namespace, but used to
have unconditional cluster scoped RBAC to do so. Combined with the fact
that it's role refers to an unrestricted PSP, users should be able to
opt out of this cluster scoped binding.

This commit does a couple of things:
- splits `podsecuritypolicy.yaml` into `role.yaml` and
  `rolebinding.yaml`, so it's more like its cluster scoped equivalents
- make both cluster scoped as namespaced role/-binding optional

Signed-off-by: Jorik Jonker <jorik.jonker@eu.equinix.com>